Add unit tests for query pagination and query sort inputs

+++ packages/amplify_datastore/android/src/main/kotlin/com/amazonaws/amplify/amplify_datastore/types/query/QueryPaginationBuilder.kt
@@ -26,12 +26,13 @@ class QueryPaginationBuilder {
                 return null
             }
             var queryPaginationInput = if (serializedMap.containsKey("page")) {
-                Page.startingAt(serializedMap["page"] as Int)
+                Page.startingAt((serializedMap["page"] as Double).toInt())
             } else {
                 Page.firstPage()
             }
             if (serializedMap.containsKey("limit")) {
-                queryPaginationInput = queryPaginationInput.withLimit(serializedMap["limit"] as Int)
+                queryPaginationInput =
+                        queryPaginationInput.withLimit((serializedMap["limit"] as Double).toInt())
             }
             return queryPaginationInput
         }
